[Feature] reorganize file structure
